+++ test/integration/bip32.js
@@ -1,13 +1,8 @@
 /* global describe, it */
 
 var assert = require('assert')
-var bigi = require('bigi')
 var bip39 = require('bip39')
 var bitcoin = require('../../')
-var crypto = require('crypto')
-
-var ecurve = require('ecurve')
-var secp256k1 = ecurve.getCurveByName('secp256k1')
 
 describe('bitcoinjs-lib (BIP32)', function () {
   it('can import a BIP32 testnet xpriv and export to WIF', function () {
@@ -17,86 +12,75 @@ describe('bitcoinjs-lib (BIP32)', function () {
     assert.equal(node.keyPair.toWIF(), 'cQfoY67cetFNunmBUX5wJiw3VNoYx3gG9U9CAofKE6BfiV1fSRw7')
   })
 
-  it('can create a BIP32 wallet external address', function () {
+  it('can export a BIP32 xpriv, then import it', function () {
+    var mnemonic = 'praise you muffin lion enable neck grocery crumble super myself license ghost'
+    var seed = bip39.mnemonicToSeed(mnemonic)
+    var node = bitcoin.HDNode.fromSeedBuffer(seed)
+    var string = node.toBase58()
+    var restored = bitcoin.HDNode.fromBase58(string)
+
+    assert.equal(node.getAddress(), restored.getAddress()) // same public key
+    assert.equal(node.keyPair.toWIF(), restored.keyPair.toWIF()) // same private key
+  })
+
+  it('can export a BIP32 xpub', function () {
+    var mnemonic = 'praise you muffin lion enable neck grocery crumble super myself license ghost'
+    var seed = bip39.mnemonicToSeed(mnemonic)
+    var node = bitcoin.HDNode.fromSeedBuffer(seed)
+    var string = node.neutered().toBase58()
+
+    assert.equal(string, 'xpub661MyMwAqRbcGhVeaVfEBA25e3cP9DsJQZoE8iep5fZSxy3TnPBNBgWnMZx56oreNc48ZoTkQfatNJ9VWnQ7ZcLZcVStpaXLTeG8bGrzX3n')
+  })
+
+  it('can create a BIP32, bitcoin, account 0, external address', function () {
     var path = "m/0'/0/0"
     var root = bitcoin.HDNode.fromSeedHex('dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd')
 
     var child1 = root.derivePath(path)
 
     // option 2, manually
-    var child2 = root.deriveHardened(0)
+    var child1b = root.deriveHardened(0)
       .derive(0)
       .derive(0)
 
     assert.equal(child1.getAddress(), '1JHyB1oPXufr4FXkfitsjgNB5yRY9jAaa7')
-    assert.equal(child2.getAddress(), '1JHyB1oPXufr4FXkfitsjgNB5yRY9jAaa7')
+    assert.equal(child1b.getAddress(), '1JHyB1oPXufr4FXkfitsjgNB5yRY9jAaa7')
   })
 
   it('can create a BIP44, bitcoin, account 0, external address', function () {
-    var path = "m/44'/0'/0'/0/0"
     var root = bitcoin.HDNode.fromSeedHex('dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd')
 
-    var child1 = root.derivePath(path)
+    var child1 = root.derivePath("m/44'/0'/0'/0/0")
 
     // option 2, manually
-    var child2 = root.deriveHardened(44)
+    var child1b = root.deriveHardened(44)
       .deriveHardened(0)
       .deriveHardened(0)
       .derive(0)
       .derive(0)
 
     assert.equal(child1.getAddress(), '12Tyvr1U8A3ped6zwMEU5M8cx3G38sP5Au')
-    assert.equal(child2.getAddress(), '12Tyvr1U8A3ped6zwMEU5M8cx3G38sP5Au')
+    assert.equal(child1b.getAddress(), '12Tyvr1U8A3ped6zwMEU5M8cx3G38sP5Au')
   })
 
-  it('can recover a BIP32 parent private key from the parent public key, and a derived, non-hardened child private key', function () {
-    function recoverParent (master, child) {
-      assert(!master.keyPair.d, 'You already have the parent private key')
-      assert(child.keyPair.d, 'Missing child private key')
+  it('can create a BIP49, bitcoin testnet, account 0, external address', function () {
+    var mnemonic = 'abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about'
+    var seed = bip39.mnemonicToSeed(mnemonic)
+    var root = bitcoin.HDNode.fromSeedBuffer(seed)
 
-      var curve = secp256k1
-      var QP = master.keyPair.Q
-      var serQP = master.keyPair.getPublicKeyBuffer()
+    var path = "m/49'/1'/0'/0/0"
+    var child = root.derivePath(path)
 
-      var d1 = child.keyPair.d
-      var d2
-      var data = Buffer.alloc(37)
-      serQP.copy(data, 0)
+    var keyhash = bitcoin.crypto.hash160(child.getPublicKeyBuffer())
+    var scriptSig = bitcoin.script.witnessPubKeyHash.output.encode(keyhash)
+    var addressBytes = bitcoin.crypto.hash160(scriptSig)
+    var outputScript = bitcoin.script.scriptHash.output.encode(addressBytes)
+    var address = bitcoin.address.fromOutputScript(outputScript, bitcoin.networks.testnet)
 
-      // search index space until we find it
-      for (var i = 0; i < bitcoin.HDNode.HIGHEST_BIT; ++i) {
-        data.writeUInt32BE(i, 33)
-
-        // calculate I
-        var I = crypto.createHmac('sha512', master.chainCode).update(data).digest()
-        var IL = I.slice(0, 32)
-        var pIL = bigi.fromBuffer(IL)
-
-        // See hdnode.js:273 to understand
-        d2 = d1.subtract(pIL).mod(curve.n)
-
-        var Qp = new bitcoin.ECPair(d2).Q
-        if (Qp.equals(QP)) break
-      }
-
-      var node = new bitcoin.HDNode(new bitcoin.ECPair(d2), master.chainCode, master.network)
-      node.depth = master.depth
-      node.index = master.index
-      node.masterFingerprint = master.masterFingerprint
-      return node
-    }
-
-    var seed = crypto.randomBytes(32)
-    var master = bitcoin.HDNode.fromSeedBuffer(seed)
-    var child = master.derive(6) // m/6
-
-    // now for the recovery
-    var neuteredMaster = master.neutered()
-    var recovered = recoverParent(neuteredMaster, child)
-    assert.strictEqual(recovered.toBase58(), master.toBase58())
+    assert.equal(address, '2Mww8dCYPUpKHofjgcXcBCEGmniw9CoaiD2')
   })
 
-  it('can use BIP39 to generate BIP32 wallet address', function () {
+  it('can use BIP39 to generate BIP32 addresses', function () {
 //     var mnemonic = bip39.generateMnemonic()
     var mnemonic = 'praise you muffin lion enable neck grocery crumble super myself license ghost'
     assert(bip39.validateMnemonic(mnemonic))
@@ -104,10 +88,12 @@ describe('bitcoinjs-lib (BIP32)', function () {
     var seed = bip39.mnemonicToSeed(mnemonic)
     var root = bitcoin.HDNode.fromSeedBuffer(seed)
 
-    // 1st receive address
+    // receive addresses
     assert.strictEqual(root.derivePath("m/0'/0/0").getAddress(), '1AVQHbGuES57wD68AJi7Gcobc3RZrfYWTC')
+    assert.strictEqual(root.derivePath("m/0'/0/1").getAddress(), '1Ad6nsmqDzbQo5a822C9bkvAfrYv9mc1JL')
 
-    // 1st change address
+    // change addresses
     assert.strictEqual(root.derivePath("m/0'/1/0").getAddress(), '1349KVc5NgedaK7DvuD4xDFxL86QN1Hvdn')
+    assert.strictEqual(root.derivePath("m/0'/1/1").getAddress(), '1EAvj4edpsWcSer3duybAd4KiR4bCJW5J6')
   })
 })