+++ test/integration/crypto.js
@@ -4,7 +4,8 @@ var assert = require('assert')
 var async = require('async')
 var bigi = require('bigi')
 var bitcoin = require('../../')
-var blockchain = require('./_blockchain')
+var mainnet = require('./_mainnet')
+var crypto = require('crypto')
 
 var ecurve = require('ecurve')
 var secp256k1 = ecurve.getCurveByName('secp256k1')
@@ -27,7 +28,7 @@ describe('bitcoinjs-lib (crypto)', function () {
     var txIds = inputs.map(function (x) { return x.txId })
 
     // first retrieve the relevant transactions
-    blockchain.m.transactions.get(txIds, function (err, results) {
+    mainnet.transactions.get(txIds, function (err, results) {
       assert.ifError(err)
 
       var transactions = {}
@@ -49,7 +50,7 @@ describe('bitcoinjs-lib (crypto)', function () {
         var prevVout = transaction.ins[input.vout].index
 
         tasks.push(function (callback) {
-          blockchain.m.transactions.get(prevOutTxId, function (err, result) {
+          mainnet.transactions.get(prevOutTxId, function (err, result) {
             if (err) return callback(err)
 
             var prevOut = bitcoin.Transaction.fromHex(result.txHex)
@@ -109,5 +110,52 @@ describe('bitcoinjs-lib (crypto)', function () {
         done()
       })
     })
+  })
+
+  it('can recover a BIP32 parent private key from the parent public key, and a derived, non-hardened child private key', function () {
+    function recoverParent (master, child) {
+      assert(!master.keyPair.d, 'You already have the parent private key')
+      assert(child.keyPair.d, 'Missing child private key')
+
+      var curve = secp256k1
+      var QP = master.keyPair.Q
+      var serQP = master.keyPair.getPublicKeyBuffer()
+
+      var d1 = child.keyPair.d
+      var d2
+      var data = Buffer.alloc(37)
+      serQP.copy(data, 0)
+
+      // search index space until we find it
+      for (var i = 0; i < bitcoin.HDNode.HIGHEST_BIT; ++i) {
+        data.writeUInt32BE(i, 33)
+
+        // calculate I
+        var I = crypto.createHmac('sha512', master.chainCode).update(data).digest()
+        var IL = I.slice(0, 32)
+        var pIL = bigi.fromBuffer(IL)
+
+        // See hdnode.js:273 to understand
+        d2 = d1.subtract(pIL).mod(curve.n)
+
+        var Qp = new bitcoin.ECPair(d2).Q
+        if (Qp.equals(QP)) break
+      }
+
+      var node = new bitcoin.HDNode(new bitcoin.ECPair(d2), master.chainCode, master.network)
+      node.depth = master.depth
+      node.index = master.index
+      node.masterFingerprint = master.masterFingerprint
+      return node
+    }
+
+    var seed = crypto.randomBytes(32)
+    var master = bitcoin.HDNode.fromSeedBuffer(seed)
+    var child = master.derive(6) // m/6
+
+    // now for the recovery
+    var neuteredMaster = master.neutered()
+    var recovered = recoverParent(neuteredMaster, child)
+    assert.strictEqual(recovered.toBase58(), master.toBase58())
   })
 })